requirements.txt
click>=8.0.0
Pillow>=9.0.0
qrcode[pil]>=7.4.2
requests>=2.25.0
# Optional: pydub>=0.25.0  # For audio conversion (may have issues with Python 3.13)
# Optional: PyPDF2>=3.0.0  # For PDF operations
# Optional: pdf2image>=1.16.0  # For PDF to image conversion 